Crystal: Limit Power Menu opacity minimum to 10
Using 0 as a value here causes unexpected behaviour from SystemUI
and will make it look like something is really broken in the
power menu.
So let's limit the minimum value to 10 which is still just barely
visible but won't cause issues with SystemUI and the Power Menu won't
seem to be completely gone.